refactor hive session into interfaces,implementation,modules
This is the bulk of the refactoring changes. `HiveSession*` objects are
converted into abstract base classes, called simply `Session*`. Interfaces are
introduced where appropriate. A new class is introduced, `SessionType`, and
it's helper `SessionTypeFactory`, allow for concrete implementation of a
`Session` to be instantiated at runtime. The new configuration
`hive.server2.session.type` is introduced to facilitate this purpose.
`Operation` instantiation is decoupled from `SessionManager`, that
responsibility is now delegated to implementations of the `OperationFactory`,
which is obtained through the `Session` implementation. The concrete
implementation of `HiveSessionType` is moved into a new submodule.